разделил написанный код на модули
