+++ js/util.js
@@ -1,0 +1,6 @@
+function checkLengthComment(str, maxlength) {
+  return str.length <= maxlength;
+}
+checkLengthComment('str', 'maxlength');
+
+export {checkLengthComment};
